remove non-linux client
