+++ client.go
@@ -1,78 +1,628 @@
+//go:build linux
+// +build linux
+
 package wifi
 
-import "fmt"
+import (
+	"bytes"
+	"crypto/sha1"
+	"fmt"
+	"net"
+	"os"
+	"time"
+	"unicode/utf8"
 
-// A Client is a type which can access WiFi device actions and statistics
-// using operating system-specific operations.
+	"github.com/mdlayher/genetlink"
+	"github.com/mdlayher/netlink"
+	"github.com/mdlayher/netlink/nlenc"
+	"golang.org/x/crypto/pbkdf2"
+	"golang.org/x/sys/unix"
+)
+
+// A client is the Linux implementation of osClient, which makes use of
+// netlink, generic netlink, and nl80211 to provide access to WiFi device
+// actions and statistics.
 type Client struct {
-	c *client
+	c             *genetlink.Conn
+	familyID      uint16
+	familyVersion uint8
 }
 
-// New creates a new Client.
-func New() (*Client, error) {
-	c, err := newClient()
+func (c *Client) String() string {
+	return fmt.Sprintf("Netlink connection: %v\nFamilyID: %v\nFamilyVersion: %v\n", c.c, c.familyID, c.familyVersion)
+}
+
+// newClient dials a generic netlink connection and verifies that nl80211
+// is available for use by this package.
+func NewClient() (*Client, error) {
+	c, err := genetlink.Dial(nil)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("NewClient: failed to create generic netlink connection: %v", err)
+	}
+
+	// Make a best effort to apply the strict options set to provide better
+	// errors and validation. We don't apply Strict in the constructor because
+	// this library is widely used on a range of kernels and we can't guarantee
+	// it will always work on older kernels.
+	for _, o := range []netlink.ConnOption{
+		netlink.ExtendedAcknowledge,
+		netlink.GetStrictCheck,
+	} {
+		_ = c.SetOption(o, true)
+	}
+
+	return initClient(c)
+}
+
+func initClient(c *genetlink.Conn) (*Client, error) {
+	family, err := c.GetFamily(unix.NL80211_GENL_NAME)
+	if err != nil {
+		// Ensure the genl socket is closed on error to avoid leaking file
+		// descriptors.
+		_ = c.Close()
+		return nil, fmt.Errorf("initClient: failed to get nl80211 family id: %v", err)
 	}
 
 	return &Client{
-		c: c,
+		c:             c,
+		familyID:      family.ID,
+		familyVersion: family.Version,
 	}, nil
 }
 
-// Close releases resources used by a Client.
-func (c *Client) Close() error {
-	return c.c.Close()
+// Close closes the client's generic netlink connection.
+func (c *Client) Close() error { return c.c.Close() }
+
+// Interfaces requests that nl80211 return a list of all WiFi interfaces present
+// on this system.
+func (c *Client) Interfaces() ([]*Interface, error) {
+	// Ask nl80211 to dump a list of all WiFi interfaces
+	msgs, err := c.get(
+		unix.NL80211_CMD_GET_INTERFACE,
+		netlink.Dump,
+		nil,
+		nil,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("Interfaces: failed to get interfaces: %v", err)
+	}
+
+	return parseInterfaces(msgs)
+}
+
+// InterfaceByName calls Interfaces() and returns one interface or nil 
+// if the requested interface doesn't exist. 
+func (c *Client) InterfaceByName(name string) (*Interface, error) {
+	interfaces, err := c.Interfaces()
+	if err != nil { return nil, fmt.Errorf("InterfaceByName: %w", err) }
+
+	for _, i := range interfaces {
+		if i.Name == name { return i, nil }
+	}
+	return nil, nil
 }
 
 // Connect starts connecting the interface to the specified ssid.
 func (c *Client) Connect(ifi *Interface, ssid string) error {
-	return c.c.Connect(ifi, ssid)
+	// Ask nl80211 to connect to the specified SSID.
+	_, err := c.get(
+		unix.NL80211_CMD_CONNECT,
+		netlink.Acknowledge,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			ae.Bytes(unix.NL80211_ATTR_SSID, []byte(ssid))
+			ae.Uint32(unix.NL80211_ATTR_AUTH_TYPE, unix.NL80211_AUTHTYPE_OPEN_SYSTEM)
+		},
+	)
+	return err
 }
 
-// Dissconnect disconnects the interface.
+// Disconnect disconnects the interface.
 func (c *Client) Disconnect(ifi *Interface) error {
-	return c.c.Disconnect(ifi)
+	// Ask nl80211 to disconnect.
+	_, err := c.get(
+		unix.NL80211_CMD_DISCONNECT,
+		netlink.Acknowledge,
+		ifi,
+		nil,
+	)
+	return err
 }
 
-// Connect starts connecting the interface to the specified ssid using WPA.
+// ConnectWPAPSK starts connecting the interface to the specified SSID using
+// WPA.
 func (c *Client) ConnectWPAPSK(ifi *Interface, ssid, psk string) error {
-	return c.c.ConnectWPAPSK(ifi, ssid, psk)
+	// Ask nl80211 to connect to the specified SSID with key..
+	_, err := c.get(
+		unix.NL80211_CMD_CONNECT,
+		netlink.Acknowledge,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			// TODO(mdlayher): document these or build from bitflags.
+			const (
+				cipherSuites = 0xfac04
+				akmSuites    = 0xfac02
+			)
+
+			ae.Bytes(unix.NL80211_ATTR_SSID, []byte(ssid))
+			ae.Uint32(unix.NL80211_ATTR_WPA_VERSIONS, unix.NL80211_WPA_VERSION_2)
+			ae.Uint32(unix.NL80211_ATTR_CIPHER_SUITE_GROUP, cipherSuites)
+			ae.Uint32(unix.NL80211_ATTR_CIPHER_SUITES_PAIRWISE, cipherSuites)
+			ae.Uint32(unix.NL80211_ATTR_AKM_SUITES, akmSuites)
+			ae.Flag(unix.NL80211_ATTR_WANT_1X_4WAY_HS, true)
+			ae.Bytes(
+				unix.NL80211_ATTR_PMK,
+				wpaPassphrase([]byte(ssid), []byte(psk)),
+			)
+			ae.Uint32(unix.NL80211_ATTR_AUTH_TYPE, unix.NL80211_AUTHTYPE_OPEN_SYSTEM)
+		},
+	)
+	return err
 }
 
-// Interfaces returns a list of the system's WiFi network interfaces.
-func (c *Client) Interfaces() ([]*Interface, error) {
-	return c.c.Interfaces()
+// wpaPassphrase computes a WPA passphrase given an SSID and preshared key.
+func wpaPassphrase(ssid, psk []byte) []byte {
+	return pbkdf2.Key(psk, ssid, 4096, 32, sha1.New)
 }
 
-// InterfaceByName calls Interfaces() and returns one interface or nil 
-// if the requested interface doesn't exist.
-func (c *Client) InterfaceByName(name string) (*Interface, error) {
-	return c.c.InterfaceByName(name)
-}
-
-// BSS retrieves the BSS associated with a WiFi interface.
+// BSS requests that nl80211 return the BSS for the specified Interface.
 func (c *Client) BSS(ifi *Interface) (*BSS, error) {
-	return c.c.BSS(ifi)
+	msgs, err := c.get(
+		unix.NL80211_CMD_GET_SCAN,
+		netlink.Dump,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			if ifi.HardwareAddr != nil {
+				ae.Bytes(unix.NL80211_ATTR_MAC, ifi.HardwareAddr)
+			}
+		},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return parseBSS(msgs)
 }
 
-// StationInfo retrieves all station statistics about a WiFi interface.
+// StationInfo requests that nl80211 return all station info for the specified
+// Interface.
 func (c *Client) StationInfo(ifi *Interface) ([]*StationInfo, error) {
-	return c.c.StationInfo(ifi)
+	msgs, err := c.get(
+		unix.NL80211_CMD_GET_STATION,
+		netlink.Dump,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			if ifi.HardwareAddr != nil {
+				ae.Bytes(unix.NL80211_ATTR_MAC, ifi.HardwareAddr)
+			}
+		},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(msgs) == 0 {
+		return nil, os.ErrNotExist
+	}
+
+	stations := make([]*StationInfo, len(msgs))
+	for i := range msgs {
+		if stations[i], err = parseStationInfo(msgs[i].Data); err != nil {
+			return nil, err
+		}
+	}
+
+	return stations, nil
 }
 
 // SetFrequency sets the frequency of a wireless interface.
+// Does nothing if frequency is already passed in value.
 func (c *Client) SetFrequency(ifi *Interface, freq int) error {
-	return c.c.SetFrequency(ifi, freq)
+	if ifi.Frequency == freq { return nil }
+	_, err := c.get(
+		unix.NL80211_CMD_SET_WIPHY,
+		netlink.Acknowledge,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			ae.Int32(unix.NL80211_ATTR_WIPHY_FREQ, int32(freq))
+		},
+	)
+	if err == nil { ifi.Frequency = freq }
+	return err
 }
 
-//SetModeMonitor sets a wireless interface iftype to MONITOR.
+// SetChannelWidth sets the channel of a wireless interface.
+func (c *Client) SetChannelWidth(ifi *Interface, width int) error {
+	_, err := c.get(
+		unix.NL80211_CMD_SET_WIPHY,
+		netlink.Acknowledge,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			ae.Int32(unix.NL80211_ATTR_CHANNEL_WIDTH, int32(width))	
+		},
+	)
+	return err
+}
+
+// SetModeMonitor sets the interface iftype to InterfaceTypeMonitor.
+// Does nothing if interface type is already InterfaceTypeMonitor.
 func (c *Client) SetModeMonitor(ifi *Interface) error {
-	fmt.Printf("Setting monitor mode on interface %v\n", ifi)
-	return c.c.SetModeMonitor(ifi)
+	if ifi.Type == InterfaceTypeMonitor { return nil }
+	_, err := c.setInterface(ifi, InterfaceTypeMonitor)
+	return err
 }
 
-//SetModeStation sets a wireless interface iftype to Station.
+// SetModeStation sets the interface iftype to InterfaceTypeStation.
+// Does nothing if the interface type is already InterfaceTypeStation.
 func (c *Client) SetModeStation(ifi *Interface) error {
-	return c.c.SetModeStation(ifi)
+	if ifi.Type == InterfaceTypeStation { return nil }
+	_, err := c.setInterface(ifi, InterfaceTypeStation)
+	return err
+}
+
+// setInterface sets the interface iftype
+func (c *Client) setInterface(ifi *Interface, iftype InterfaceType) ([]genetlink.Message, error) {
+	msgs, err := c.get(
+		unix.NL80211_CMD_SET_INTERFACE,
+		netlink.Acknowledge,
+		ifi,
+		func(ae *netlink.AttributeEncoder) {
+			ae.Int32(unix.NL80211_ATTR_IFTYPE, int32(iftype))
+		},
+	)
+	if err != nil {
+		return nil, err
+	}
+	ifi.Type = iftype
+	if len(msgs) == 0 {
+		return nil, os.ErrNotExist
+	}
+	return msgs, nil	
+}
+
+// get performs a request/response interaction with nl80211.
+func (c *Client) get(
+	cmd uint8,
+	flags netlink.HeaderFlags,
+	ifi *Interface,
+	// May be nil; used to apply optional parameters.
+	params func(ae *netlink.AttributeEncoder),
+) ([]genetlink.Message, error) {
+	ae := netlink.NewAttributeEncoder()
+	ifi.encode(ae)
+	if params != nil {
+		// Optionally apply more parameters to the attribute encoder.
+		params(ae)
+	}
+
+	// Note: don't send netlink.Acknowledge or we get an extra message back from
+	// the kernel which doesn't seem useful as of now.
+	return c.execute(cmd, flags, ae)
+}
+
+// execute executes the specified command with additional header flags and input
+// netlink request attributes. The netlink.Request header flag is automatically
+// set.
+func (c *Client) execute(
+	cmd uint8,
+	flags netlink.HeaderFlags,
+	ae *netlink.AttributeEncoder,
+) ([]genetlink.Message, error) {
+	b, err := ae.Encode()
+	if err != nil {
+		return nil, err
+	}
+
+	return c.c.Execute(
+		genetlink.Message{
+			Header: genetlink.Header{
+				Command: cmd,
+				Version: c.familyVersion,
+			},
+			Data: b,
+		},
+		// Always pass the genetlink family ID and request flag.
+		c.familyID,
+		netlink.Request|flags,
+	)
+}
+
+// parseInterfaces parses zero or more Interfaces from nl80211 interface
+// messages.
+func parseInterfaces(msgs []genetlink.Message) ([]*Interface, error) {
+	ifis := make([]*Interface, 0, len(msgs))
+	for _, m := range msgs {
+		attrs, err := netlink.UnmarshalAttributes(m.Data)
+		if err != nil {
+			return nil, err
+		}
+
+		var ifi Interface
+		if err := (&ifi).parseAttributes(attrs); err != nil {
+			return nil, err
+		}
+
+		ifis = append(ifis, &ifi)
+	}
+
+	return ifis, nil
+}
+
+// encode provides an encoding function for ifi's attributes. If ifi is nil,
+// encode is a no-op.
+func (ifi *Interface) encode(ae *netlink.AttributeEncoder) {
+	if ifi == nil {
+		return
+	}
+
+	// Mandatory.
+	ae.Uint32(unix.NL80211_ATTR_IFINDEX, uint32(ifi.Index))
+}
+
+// idAttrs returns the netlink attributes required from an Interface to retrieve
+// more data about it.
+func (ifi *Interface) idAttrs() []netlink.Attribute {
+	return []netlink.Attribute{
+		{
+			Type: unix.NL80211_ATTR_IFINDEX,
+			Data: nlenc.Uint32Bytes(uint32(ifi.Index)),
+		},
+		{
+			Type: unix.NL80211_ATTR_MAC,
+			Data: ifi.HardwareAddr,
+		},
+	}
+}
+
+// parseAttributes parses netlink attributes into an Interface's fields.
+func (ifi *Interface) parseAttributes(attrs []netlink.Attribute) error {
+	for _, a := range attrs {
+		switch a.Type {
+		case unix.NL80211_ATTR_IFINDEX:
+			ifi.Index = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_ATTR_IFNAME:
+			ifi.Name = nlenc.String(a.Data)
+		case unix.NL80211_ATTR_MAC:
+			ifi.HardwareAddr = net.HardwareAddr(a.Data)
+		case unix.NL80211_ATTR_WIPHY:
+			ifi.PHY = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_ATTR_IFTYPE:
+			// NOTE: InterfaceType copies the ordering of nl80211's interface type
+			// constants.  This may not be the case on other operating systems.
+			ifi.Type = InterfaceType(nlenc.Uint32(a.Data))
+		case unix.NL80211_ATTR_WDEV:
+			ifi.Device = int(nlenc.Uint64(a.Data))
+		case unix.NL80211_ATTR_WIPHY_FREQ:
+			ifi.Frequency = int(nlenc.Uint32(a.Data))
+		}
+	}
+
+	return nil
+}
+
+// parseBSS parses a single BSS with a status attribute from nl80211 BSS messages.
+func parseBSS(msgs []genetlink.Message) (*BSS, error) {
+	for _, m := range msgs {
+		attrs, err := netlink.UnmarshalAttributes(m.Data)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, a := range attrs {
+			if a.Type != unix.NL80211_ATTR_BSS {
+				continue
+			}
+
+			nattrs, err := netlink.UnmarshalAttributes(a.Data)
+			if err != nil {
+				return nil, err
+			}
+
+			// The BSS which is associated with an interface will have a status
+			// attribute
+			if !attrsContain(nattrs, unix.NL80211_BSS_STATUS) {
+				continue
+			}
+
+			var bss BSS
+			if err := (&bss).parseAttributes(nattrs); err != nil {
+				return nil, err
+			}
+
+			return &bss, nil
+		}
+	}
+
+	return nil, os.ErrNotExist
+}
+
+// parseAttributes parses netlink attributes into a BSS's fields.
+func (b *BSS) parseAttributes(attrs []netlink.Attribute) error {
+	for _, a := range attrs {
+		switch a.Type {
+		case unix.NL80211_BSS_BSSID:
+			b.BSSID = net.HardwareAddr(a.Data)
+		case unix.NL80211_BSS_FREQUENCY:
+			b.Frequency = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_BSS_BEACON_INTERVAL:
+			// Raw value is in "Time Units (TU)".  See:
+			// https://en.wikipedia.org/wiki/Beacon_frame
+			b.BeaconInterval = time.Duration(nlenc.Uint16(a.Data)) * 1024 * time.Microsecond
+		case unix.NL80211_BSS_SEEN_MS_AGO:
+			// * @NL80211_BSS_SEEN_MS_AGO: age of this BSS entry in ms
+			b.LastSeen = time.Duration(nlenc.Uint32(a.Data)) * time.Millisecond
+		case unix.NL80211_BSS_STATUS:
+			// NOTE: BSSStatus copies the ordering of nl80211's BSS status
+			// constants.  This may not be the case on other operating systems.
+			b.Status = BSSStatus(nlenc.Uint32(a.Data))
+		case unix.NL80211_BSS_INFORMATION_ELEMENTS:
+			ies, err := parseIEs(a.Data)
+			if err != nil {
+				return err
+			}
+
+			// TODO(mdlayher): return more IEs if they end up being generally useful
+			for _, ie := range ies {
+				switch ie.ID {
+				case ieSSID:
+					b.SSID = decodeSSID(ie.Data)
+				}
+			}
+		}
+	}
+
+	return nil
+}
+
+// parseStationInfo parses StationInfo attributes from a byte slice of
+// netlink attributes.
+func parseStationInfo(b []byte) (*StationInfo, error) {
+	attrs, err := netlink.UnmarshalAttributes(b)
+	if err != nil {
+		return nil, err
+	}
+
+	var info StationInfo
+	for _, a := range attrs {
+		switch a.Type {
+		case unix.NL80211_ATTR_MAC:
+			info.HardwareAddr = net.HardwareAddr(a.Data)
+		case unix.NL80211_ATTR_STA_INFO:
+			nattrs, err := netlink.UnmarshalAttributes(a.Data)
+			if err != nil {
+				return nil, err
+			}
+
+			if err := (&info).parseAttributes(nattrs); err != nil {
+				return nil, err
+			}
+
+			// Parsed the necessary data.
+			return &info, nil
+		}
+	}
+
+	// No station info found
+	return nil, os.ErrNotExist
+}
+
+// parseAttributes parses netlink attributes into a StationInfo's fields.
+func (info *StationInfo) parseAttributes(attrs []netlink.Attribute) error {
+	for _, a := range attrs {
+		switch a.Type {
+		case unix.NL80211_STA_INFO_CONNECTED_TIME:
+			// Though nl80211 does not specify, this value appears to be in seconds:
+			// * @NL80211_STA_INFO_CONNECTED_TIME: time since the station is last connected
+			info.Connected = time.Duration(nlenc.Uint32(a.Data)) * time.Second
+		case unix.NL80211_STA_INFO_INACTIVE_TIME:
+			// * @NL80211_STA_INFO_INACTIVE_TIME: time since last activity (u32, msecs)
+			info.Inactive = time.Duration(nlenc.Uint32(a.Data)) * time.Millisecond
+		case unix.NL80211_STA_INFO_RX_BYTES64:
+			info.ReceivedBytes = int(nlenc.Uint64(a.Data))
+		case unix.NL80211_STA_INFO_TX_BYTES64:
+			info.TransmittedBytes = int(nlenc.Uint64(a.Data))
+		case unix.NL80211_STA_INFO_SIGNAL:
+			//  * @NL80211_STA_INFO_SIGNAL: signal strength of last received PPDU (u8, dBm)
+			// Should just be cast to int8, see code here: https://git.kernel.org/pub/scm/linux/kernel/git/jberg/iw.git/tree/station.c#n378
+			info.Signal = int(int8(a.Data[0]))
+		case unix.NL80211_STA_INFO_RX_PACKETS:
+			info.ReceivedPackets = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_STA_INFO_TX_PACKETS:
+			info.TransmittedPackets = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_STA_INFO_TX_RETRIES:
+			info.TransmitRetries = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_STA_INFO_TX_FAILED:
+			info.TransmitFailed = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_STA_INFO_BEACON_LOSS:
+			info.BeaconLoss = int(nlenc.Uint32(a.Data))
+		case unix.NL80211_STA_INFO_RX_BITRATE, unix.NL80211_STA_INFO_TX_BITRATE:
+			rate, err := parseRateInfo(a.Data)
+			if err != nil {
+				return err
+			}
+
+			// TODO(mdlayher): return more statistics if they end up being
+			// generally useful
+			switch a.Type {
+			case unix.NL80211_STA_INFO_RX_BITRATE:
+				info.ReceiveBitrate = rate.Bitrate
+			case unix.NL80211_STA_INFO_TX_BITRATE:
+				info.TransmitBitrate = rate.Bitrate
+			}
+		}
+
+		// Only use 32-bit counters if the 64-bit counters are not present.
+		// If the 64-bit counters appear later in the slice, they will overwrite
+		// these values.
+		if info.ReceivedBytes == 0 && a.Type == unix.NL80211_STA_INFO_RX_BYTES {
+			info.ReceivedBytes = int(nlenc.Uint32(a.Data))
+		}
+		if info.TransmittedBytes == 0 && a.Type == unix.NL80211_STA_INFO_TX_BYTES {
+			info.TransmittedBytes = int(nlenc.Uint32(a.Data))
+		}
+	}
+
+	return nil
+}
+
+// rateInfo provides statistics about the receive or transmit rate of
+// an interface.
+type rateInfo struct {
+	// Bitrate in bits per second.
+	Bitrate int
+}
+
+// parseRateInfo parses a rateInfo from netlink attributes.
+func parseRateInfo(b []byte) (*rateInfo, error) {
+	attrs, err := netlink.UnmarshalAttributes(b)
+	if err != nil {
+		return nil, err
+	}
+
+	var info rateInfo
+	for _, a := range attrs {
+		switch a.Type {
+		case unix.NL80211_RATE_INFO_BITRATE32:
+			info.Bitrate = int(nlenc.Uint32(a.Data))
+		}
+
+		// Only use 16-bit counters if the 32-bit counters are not present.
+		// If the 32-bit counters appear later in the slice, they will overwrite
+		// these values.
+		if info.Bitrate == 0 && a.Type == unix.NL80211_RATE_INFO_BITRATE {
+			info.Bitrate = int(nlenc.Uint16(a.Data))
+		}
+	}
+
+	// Scale bitrate to bits/second as base unit instead of 100kbits/second.
+	// * @NL80211_RATE_INFO_BITRATE: total bitrate (u16, 100kbit/s)
+	info.Bitrate *= 100 * 1000
+
+	return &info, nil
+}
+
+// attrsContain checks if a slice of netlink attributes contains an attribute
+// with the specified type.
+func attrsContain(attrs []netlink.Attribute, typ uint16) bool {
+	for _, a := range attrs {
+		if a.Type == typ {
+			return true
+		}
+	}
+
+	return false
+}
+
+// decodeSSID safely parses a byte slice into UTF-8 runes, and returns the
+// resulting string from the runes.
+func decodeSSID(b []byte) string {
+	buf := bytes.NewBuffer(nil)
+	for len(b) > 0 {
+		r, size := utf8.DecodeRune(b)
+		b = b[size:]
+
+		buf.WriteRune(r)
+	}
+
+	return buf.String()
 }
